Fix infinite orders/trades WS topic resubscription in UI
